server/weave-sidecar/requirements.txt
weave
wandb
fastapi
uvicorn[standard]
pydantic>=2.0
python-dotenv
